% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetBCDataBatch.R
\name{get_bcdata_batch}
\alias{get_bcdata_batch}
\title{Download BC Data layers in batch.}
\usage{
get_bcdata_batch(layer_tbl)
}
\arguments{
\item{layer_tbl}{A data frame with three columns named 'code','geom_ext'
& 'filter_exp'. The 'code' field represents a character string denoting
the name or ID of a resource or the URL or a BC Geographic Warehouse BCGW
name. The 'geom_ext' represents the name of the SF object as a string for
which to intersect the layer data with, which must be present in the global
environment. The 'filter_exp' field represents the filter expression as a
string to be passed to bcdata::filter evaluated by bcdata::CQL.}
}
\value{
A list of the resulting BC Data layers defined by 'layer_tbl'.
}
\description{
This function downloads layers from the BC Geographic Warehouse
using the R bcdata package. The function takes in one input
as a data frame with three columns named 'code','geom_ext'
& 'filter_exp'. Each row represents a new BC Data layer,
and its associated parameters. See table parameter
descriptions below.
}
